+++ spec/support/table_loader.rb
@@ -14,6 +14,7 @@ module TableLoader
     transactions
     reviews
     system_announcements
+    posts
   ].each do |table_name|
     define_method("#{table_name}_table") do |adapter|
       adapter = adapter.to_sym